Fix floating point issues by storing and querying on Decimal128 types

+++ Sanctuary.Census.Common/Objects/Collections/Experience.cs
@@ -18,6 +18,6 @@ public record Experience
     [property:Key] uint ExperienceID,
     uint AwardTypeID,
     LocaleString Description,
-    float XP,
+    decimal XP,
     bool IsNotableEvent
 );
